Starting tests for closable
