+++ contracts/Phase/Closable.sol
@@ -4,6 +4,7 @@ contract Closable {
     uint256 public closingTime;
 
     function Closable(uint256 _closingTime) public {
+        require(_closingTime > now);
         closingTime = _closingTime;
     }
